🔨 Refactor(#85): useCosmeticItemStore 경로 수정 및 불필요한 악세사리 옵션 제거
